pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt<4.1",
    "faker>=37.12.0",
    "fastapi>=0.120.4",
    "google-genai>=1.47.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
    "uvicorn[standard]>=0.38.0",
]
